+++ frontend/src/lib/components/StringTypeNarrowing.svelte
@@ -1,12 +1,15 @@
 <script lang="ts">
 	import RadioButton from './RadioButton.svelte'
 	import ResourceTypePicker from './ResourceTypePicker.svelte'
+	import Toggle from './Toggle.svelte'
 	import { Button } from './common'
+	import RegexGen from './copilot/RegexGen.svelte'
 
 	export let pattern: string | undefined
 	export let enum_: string[] | undefined
 	export let format: string | undefined
 	export let contentEncoding: 'base64' | 'binary' | undefined
+	export let customErrorMessage: string | undefined
 
 	let kind: 'none' | 'pattern' | 'enum' | 'resource' | 'format' | 'base64' = computeKind()
 	let patternStr: string = pattern ?? ''
@@ -32,6 +35,11 @@
 	$: pattern = patternStr == '' ? undefined : patternStr
 	$: contentEncoding = kind == 'base64' ? 'base64' : undefined
 
+	$: {
+		if (format == 'email') {
+			pattern = '^[\\w-.]+@([\\w-]+\\.)+[\\w-]{2,4}$'
+		}
+	}
 	function add() {
 		enum_ = enum_ ? enum_.concat('') : ['']
 	}
@@ -73,18 +81,30 @@
 		['Pattern', 'pattern']
 	]}
 	bind:value={kind}
+	on:change={(e) => {
+		if (e.detail != 'enum') {
+			enum_ = undefined
+		}
+	}}
 />
 <div class="my-2" />
 
 {#if kind == 'pattern'}
 	<label for="input" class="mb-2 text-secondary text-xs">
 		Pattern (Regex)
-		<div class="flex flex-row">
+		<div class="flex flex-row items-center gap-0.5">
 			<input
 				id="input"
 				type="text"
 				placeholder="^(\\([0-9]{3}\\))?[0-9]{3}-[0-9]{4}$"
 				bind:value={patternStr}
+			/>
+			<RegexGen
+				on:gen={(e) => {
+					const { res, prompt } = e.detail
+					patternStr = res
+					customErrorMessage = 'does not match: ' + prompt
+				}}
 			/>
 			<Button
 				variant="border"
@@ -97,6 +117,22 @@
 			>
 				clear
 			</Button>
+		</div>
+		<div class="mt-2 flex gap-2">
+			<Toggle
+				size="xs"
+				options={{ right: 'Custom error message' }}
+				checked={customErrorMessage != undefined && customErrorMessage != ''}
+				on:change={(e) => {
+					if (e.detail) {
+						customErrorMessage = 'Custom error message'
+					} else {
+						customErrorMessage = undefined
+					}
+				}}
+				>Custom error message
+			</Toggle>
+			<input type="text" bind:value={customErrorMessage} />
 		</div>
 	</label>
 {:else if kind == 'enum'}