+++ frontend/src/lib/components/common/skeleton/Skeleton.svelte
@@ -1,20 +1,26 @@
 <script lang="ts">
-	import { fade } from 'svelte/transition'
 	import { HEIGHT_UNIT, type SkeletonLayout } from './model'
 	import SkeletonElement from './SkeletonElement.svelte'
+	import { onMount } from 'svelte'
 
 	export let layout: SkeletonLayout
 	export let loading = true
 	export let overlay = false
+
+	export let mounted = false
+	onMount(() => {
+		setTimeout(() => {
+			mounted = true
+		}, 10)
+	})
 </script>
 
 {#if loading}
 	<div class="relative flex justify-center">
 		<div
-			in:fade|local={{ duration: 1000 }}
-			class="flex grow flex-col overflow-hidden {overlay
-				? 'absolute w-full h-full z-[1000]'
-				: ''} {$$props.class}"
+			class="flex grow flex-col overflow-hidden transition-[opacity] duration-1000 opacity-0 {mounted
+				? 'opacity-100'
+				: ''} {overlay ? 'absolute w-full h-full z-[1000]' : ''} {$$props.class}"
 		>
 			{#each layout as row}
 				<div class="flex justify-between items-start gap-4">